seperated out tests for ai into their own classes. Added 2 more boards for the ai to test. fixed a minor bug with ai.generateMove
