+++ src/ai/Ai.java
@@ -12,7 +12,6 @@ import move_commands.MoveCommand;
 import move_commands.PullMove;
 import move_commands.PushMove;
 import move_commands.RegularMove;
-import piece.AbstractPiece;
 import piece.Owner;
 
 public class Ai {
@@ -20,21 +19,12 @@ public class Ai {
 	public static final long AVERAGE_TIME_LIMIT = 500_000;// ns, 0.5 miliseconds
 	private Owner owner;
 	private Game game;
-	private int times;
 
 	enum MoveType {
 		Regular, Push, Pull;
 		static MoveType getRandomMoveType() {
 			return MoveType.values()[MoveType.values().length];
 		}
-	}
-
-	interface returnsCoordinate {
-		public Coordinate execute();
-	}
-
-	interface returnsAbstractPiece {
-		public AbstractPiece execute();
 	}
 
 	public Ai(Owner owner, Game game) {
@@ -56,9 +46,7 @@ public class Ai {
 	 * @return
 	 */
 	public MoveCommand generateMove() {
-		this.times++;
 		Coordinate pieceCoor = this.generateRandomPieceCoor();
-		AbstractPiece piece = this.game.getPieceAt(pieceCoor);
 		Coordinate randomDirection = this.generateRandomDirection(pieceCoor);
 		if (!randomDirection.isValid()) {
 			// can't move here because its off the board, try again
@@ -100,7 +88,11 @@ public class Ai {
 		}
 		// use regular method
 		MoveCommand regularMove = new RegularMove(this.game.getBoardState(), pieceCoor, randomDirection, this.owner);
-		return regularMove;
+		if (regularMove.isValidMove()) {
+			return regularMove;
+		}
+		// invalid regular move, try again
+		return this.generateMove();
 	}
 
 	// 1 / 4 chance of generating pull, 3 possible pull moves, and 1 regular move
